+++ data-protocols/ids/ids-api-multipart-endpoint-v1/src/test/java/org/eclipse/dataspaceconnector/ids/api/multipart/IdsApiMultipartEndpointV1IntegrationTestServiceExtension.java
@@ -304,11 +304,6 @@ class IdsApiMultipartEndpointV1IntegrationTestServiceExtension implements Servic
         public void reload() {
             throw new NotImplementedError();
         }
-
-        @Override
-        public Stream<ContractDefinition> isReferenced(String policyId) {
-            throw new UnsupportedOperationException();
-        }
     }
 
     private static class FakeContractValidationService implements ContractValidationService {