+++ extensions/sql/contract-definition-store-sql/docs/schema.sql
@@ -19,6 +19,6 @@ CREATE TABLE IF NOT EXISTS edc_contract_definitions
     contract_definition_id VARCHAR NOT NULL,
     access_policy_id       VARCHAR NOT NULL,
     contract_policy_id     VARCHAR NOT NULL,
-    selector_expression    VARCHAR NOT NULL,
+    selector_expression    JSON NOT NULL,
     PRIMARY KEY (contract_definition_id)
 );