+++ CHANGELOG.md
@@ -25,7 +25,8 @@ in the detailed section referring to by linking pull requests or issues.
 #### Changed
 
 * Provided default no-op `TransactionContext` (#1461)
-* Refactored query capabilites for `Asset` (#1459)
+* Refactored query capabilities for `Asset` (#1459)
+* Refactored query capabilities for `ContractDefinition` (#1458)
 
 #### Removed